refactor: MapView 전체 리팩토링
